requirements.txt
pandas
psycopg2-binary
openai
python-dotenv
